pyproject.toml
[tool.poetry]
name = "transformers"
version = "0.1.0"
description = "Exercises in building useful data transformation utilities."
authors = ["Department for International Trade"]
license = "[provisionally] BSD"

[tool.poetry.dependencies]
python = "^3.6,^3.7,^3.8,^3.9"
mongoengine = "^0.24.0"
hu = "^0.9.0"
google-auth = "^2.6.2"
google_api = "^0.1.12"
requests = "^2.27.1"
google-auth-oauthlib = "^0.5.1"
google-api-python-client = "^2.41.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
